// src/enums/PianoNotes.ts
export enum PianoNotes {
  C = "C",
  CSharp = "C#",
  DFlat = "Db",
  D = "D",
  DSharp = "D#",
  EFlat = "Eb",
  E = "E",
  F = "F",
  FSharp = "F#",
  GFlat = "Gb",
  G = "G",
  GSharp = "G#",
  AFlat = "Ab",
  A = "A",
  ASharp = "A#",
  BFlat = "Bb",
  B = "B",
  C2 = "C2",
  CSharp2 = "C#2",
  DFlat2 = "Db2",
  D2 = "D2",
  DSharp2 = "D#2",
  EFlat2 = "Eb2",
  E2 = "E2",
  F2 = "F2",
  FSharp2 = "F#2",
  GFlat2 = "Gb2",
  G2 = "G2",
  GSharp2 = "G#2",
  AFlat2 = "Ab2",
  A2 = "A2",
  ASharp2 = "A#2",
  BFlat2 = "Bb2",
  B2 = "B2",
}